+++ src/main/java/http/request/exception/InvalidRequestException.java
@@ -1,4 +1,4 @@
-package http.request;
+package http.request.exception;
 
 public class InvalidRequestException extends RuntimeException {
     private static final String INVALID_REQUEST_EXCEPTION = "잘못된 HTTP 요청입니다.";
